+++ Libraries/LibWeb/CMakeLists.txt
@@ -654,6 +654,7 @@ set(SOURCES
     HTML/TraversableNavigable.cpp
     HTML/UniversalGlobalScope.cpp
     HTML/UserActivation.cpp
+    HTML/UserNavigationInvolvement.cpp
     HTML/ValidityState.cpp
     HTML/VideoTrack.cpp
     HTML/VideoTrackList.cpp
