+++ Libraries/LibWeb/CSS/StyleValues/StyleValue.cpp
@@ -74,6 +74,7 @@
 #include <LibWeb/CSS/StyleValues/URLStyleValue.h>
 #include <LibWeb/CSS/StyleValues/UnicodeRangeStyleValue.h>
 #include <LibWeb/CSS/StyleValues/UnresolvedStyleValue.h>
+#include <LibWeb/HTML/Navigable.h>
 #include <LibWeb/Layout/Node.h>
 
 namespace Web::CSS {
