+++ Libraries/LibWeb/CSS/VisualViewport.cpp
@@ -11,6 +11,7 @@
 #include <LibWeb/DOM/Document.h>
 #include <LibWeb/DOM/EventDispatcher.h>
 #include <LibWeb/HTML/EventNames.h>
+#include <LibWeb/HTML/Navigable.h>
 
 namespace Web::CSS {
