+++ Libraries/LibWeb/Forward.h
@@ -781,6 +781,7 @@ struct EnvironmentSettingsObject;
 struct NavigationParams;
 struct OpenerPolicy;
 struct OpenerPolicyEnforcementResult;
+struct PaintConfig;
 struct PolicyContainer;
 struct POSTResource;
 struct ScrollOptions;
@@ -788,6 +789,7 @@ struct ScrollToOptions;
 struct SerializedFormData;
 struct SerializedPolicyContainer;
 struct SerializedTransferRecord;
+struct SourceSnapshotParams;
 struct StructuredSerializeOptions;
 struct SyntheticRealmSettings;
 struct ToggleTaskTracker;
