+++ Libraries/LibWeb/HTML/CrossOrigin/Reporting.cpp
@@ -12,6 +12,7 @@
 #include <LibWeb/HTML/BrowsingContext.h>
 #include <LibWeb/HTML/CrossOrigin/AbstractOperations.h>
 #include <LibWeb/HTML/CrossOrigin/Reporting.h>
+#include <LibWeb/HTML/Navigable.h>
 
 namespace Web::HTML {
