+++ Libraries/LibWeb/HTML/InitialInsertion.h
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2025, Sam Atkins <sam@ladybird.org>
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+namespace Web::HTML {
+
+enum class InitialInsertion : u8 {
+    Yes,
+    No,
+};
+
+}
