+++ Libraries/LibWeb/HTML/NavigableContainer.h
@@ -8,7 +8,7 @@
 
 #include <LibWeb/Export.h>
 #include <LibWeb/HTML/HTMLElement.h>
-#include <LibWeb/HTML/Navigable.h>
+#include <LibWeb/HTML/InitialInsertion.h>
 
 namespace Web::HTML {
 
@@ -39,12 +39,7 @@ public:
     // https://html.spec.whatwg.org/multipage/iframe-embed-object.html#potentially-delays-the-load-event
     bool currently_delays_the_load_event() const;
 
-    bool content_navigable_has_session_history_entry_and_ready_for_navigation() const
-    {
-        if (!content_navigable())
-            return false;
-        return m_content_navigable->has_session_history_entry_and_ready_for_navigation();
-    }
+    bool content_navigable_has_session_history_entry_and_ready_for_navigation() const;
 
 protected:
     NavigableContainer(DOM::Document&, DOM::QualifiedName);
@@ -64,12 +59,7 @@ protected:
 
     void set_potentially_delays_the_load_event(bool value) { m_potentially_delays_the_load_event = value; }
 
-    void set_content_navigable_has_session_history_entry_and_ready_for_navigation()
-    {
-        if (!content_navigable())
-            return;
-        content_navigable()->set_has_session_history_entry_and_ready_for_navigation();
-    }
+    void set_content_navigable_has_session_history_entry_and_ready_for_navigation();
 
 private:
     virtual bool is_navigable_container() const override { return true; }
