+++ Libraries/LibWeb/MixedContent/AbstractOperations.cpp
@@ -5,6 +5,7 @@
  */
 
 #include <LibWeb/Fetch/Response.h>
+#include <LibWeb/HTML/Navigable.h>
 #include <LibWeb/HTML/Window.h>
 #include <LibWeb/MixedContent/AbstractOperations.h>
 #include <LibWeb/SecureContexts/AbstractOperations.h>
