+++ Libraries/LibWeb/Painting/NavigableContainerViewportPaintable.cpp
@@ -5,6 +5,7 @@
  */
 
 #include <AK/Debug.h>
+#include <LibWeb/HTML/Navigable.h>
 #include <LibWeb/HTML/NavigableContainer.h>
 #include <LibWeb/Layout/NavigableContainerViewport.h>
 #include <LibWeb/Layout/Viewport.h>
