+++ Libraries/LibWeb/SVG/SVGScriptElement.cpp
@@ -8,6 +8,7 @@
 #include <LibWeb/Bindings/SVGScriptElementPrototype.h>
 #include <LibWeb/Fetch/Fetching/Fetching.h>
 #include <LibWeb/Fetch/Infrastructure/FetchAlgorithms.h>
+#include <LibWeb/Fetch/Infrastructure/HTTP/Responses.h>
 #include <LibWeb/HTML/Scripting/ClassicScript.h>
 #include <LibWeb/Namespace.h>
 #include <LibWeb/SVG/AttributeNames.h>
